refactor: change directory of test classes
