+++ src/test/kotlin/lotto/model/WinningLottoTest.kt
@@ -1,6 +1,6 @@
-package lotto
+package lotto.model
 
-import lotto.model.WinningLotto
+import lotto.Lotto
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
